+++ components/Drawer/types.ts
@@ -1,7 +1,8 @@
 export interface IDrawerPropTypes extends React.PropsWithChildren {
   className?: string;
-  title?: React.ReactNode;
+  maskClosable?: boolean;
   onClose?: () => void;
   open?: boolean;
+  title?: React.ReactNode;
   width?: number | string;
 }