+++ hooks/index.ts
@@ -1,3 +1,4 @@
 export * from "./useChannel";
 export * from "./useStream";
 export * from "./useOpenAIKey";
+export * from "./useProxy";